Build libgpg-error and libgcrypt under linux, for OTR.
There is still a problem, the linux compile script (number 3) will bomb out at the link stage, but simply starting it again it runs fine.
This was too much work, and it's way past my bed time.  It works, so commiting it now, and fixing that problem later.

Still could use some cleaning up to.