Next part of making the meshes real.
There's a nasty runaway memory leak that locks up the viewer in here somewhere.
Though I suspect it's coz something is not being done right, and triggering it in code I've not looked at yet.

The textures are not added to the meshes yet, and they end up being invisible, even the prims linked to it.
You can select the meshes though, but they are messed up.

Somewhere between creating all this data and the other end of the evil pipeline code, things get corrupted.
The solution is buried deep in that pipeline somewhere, or I'm not doing some step right.

Note that the existing code expects quads, and we only have triangles, that's likely one source of problems.