Revert "Another possible fix for misplaced HUDs?"
This reverts commit a1f82acebab97347d49a941f67a6df0ce78c29a6.

It screws with TPs, and does not seem to help anyway.

Note that this is really a server bug we keep trying to work around in the viewer.  We really should just fix the damn server.  It's the server that keeps things laying around in it's cache that are no longer there, then sending them to the viewer and confusing it.