Support the new WindlightReset message from OpenSim for resetting Windlight presets when leaving a sim that uses Lightshare.
Note, this windlight and lighthshare stuff needs a refactoring.
The (I think LL inspired) Windlight CAPS duplicates a lot of this code.
Still should backup any WL changes that have not been stored in a preset to.
Some cleanup to do as well.
That's why there is some excess cruft added with this commit, part way through implementing some of the above.